% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graffle-package.R, R/graffle.R
\docType{package}
\name{graffle}
\alias{graffle}
\alias{graffle-package}
\title{An 'OmniGraffle' Graphics Device}
\usage{
graffle(width = 800, height = 600, bg = "white", pointsize = 12,
  res = 72, clip = FALSE, antialias = FALSE, system_fonts = list(),
  user_fonts = list())
}
\arguments{
\item{width}{width}

\item{height}{height}

\item{bg}{bg}

\item{pointsize}{pointsize}

\item{res}{res}

\item{clip}{clip}

\item{antialias}{antialias}

\item{system_fonts}{system_fonts}

\item{user_fonts}{user_fonts}
}
\description{
A graphics device that produces 'OmniGraffle' documents.
Currently based on the code structure of the 'svglite' package.

NOTE: OmniGraffle needs to be open and have a document open.
}
\examples{
\dontrun{
graffle()

}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
